Create README.md
This system manages to the analysis of the creation and develops manual entry of the player information like profile fees attendance and tour. The main aim of Cricket Management is to manage the details of Cricket, Players, Attendance, fees. It manages all the information about Cricket, Fees, tour. The project is totally built at administrative end and thus only the administrator is guaranteed the access. The purpose of the project is to build an application program to reduce the manual work for managing the Cricket, Fees, Players, attendances and fees. It tracks all the details about the Players, Attendance, fees and tour.